Fix : reduce chatfeedrecycleradapter complexity
Fix : reduce chatfeedrecycleradapter complexity

Fix : reduce chatfeedrecycleradapter complexity

Fix : reduce chatfeedrecycleradapter complexity

Fix : reduce chatfeedrecycleradapter complexity

Fix : reduce chatfeedrecycleradapter complexity

Fix : reduce chatfeedrecycleradapter complexity

Fix : reduce complexity of chatfeedrecycleradapter

Fix : reduce complexity of chatfeedrecycleradapter

Fix : reduce complexity of chatfeedrecycleradapter

Fix : reduce complexity of chatfeedrecycleradapter

Fix : reduce chatfeedrecycleradapter complexity

Fix : reduce chatfeedrecycleradapter complexity

Fix : reduce chatfeedrecycleradapter complexity

Fix : reduce chatfeedrecycleradapter complexity
