+++ app/src/main/java/org/fossasia/susi/ai/adapters/recycleradapters/ChatFeedRecyclerAdapter.java
@@ -2,22 +2,13 @@ package org.fossasia.susi.ai.adapters.recycleradapters;
 
 import android.content.Context;
 import android.content.Intent;
-import android.graphics.Color;
-import android.net.Uri;
 import android.os.Build;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.support.v4.content.ContextCompat;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.view.ActionMode;
-import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
-import android.text.Html;
-import android.text.Spannable;
-import android.text.SpannableString;
-import android.text.Spanned;
-import android.text.method.LinkMovementMethod;
-import android.text.style.BackgroundColorSpan;
 import android.util.Log;
 import android.util.Patterns;
 import android.util.SparseBooleanArray;
@@ -29,16 +20,6 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Toast;
 
-import com.github.mikephil.charting.data.Entry;
-import com.github.mikephil.charting.data.PieData;
-import com.github.mikephil.charting.data.PieDataSet;
-import com.github.mikephil.charting.formatter.PercentFormatter;
-import com.github.mikephil.charting.utils.ColorTemplate;
-import com.leocardz.link.preview.library.LinkPreviewCallback;
-import com.leocardz.link.preview.library.SourceContent;
-import com.leocardz.link.preview.library.TextCrawler;
-import com.squareup.picasso.Picasso;
-
 import org.fossasia.susi.ai.R;
 import org.fossasia.susi.ai.adapters.viewholders.ChatViewHolder;
 import org.fossasia.susi.ai.adapters.viewholders.DateViewHolder;
@@ -49,34 +30,19 @@ import org.fossasia.susi.ai.adapters.viewholders.PieChartViewHolder;
 import org.fossasia.susi.ai.adapters.viewholders.SearchResultsListHolder;
 import org.fossasia.susi.ai.adapters.viewholders.TypingDotsHolder;
 import org.fossasia.susi.ai.adapters.viewholders.ZeroHeightHolder;
-import org.fossasia.susi.ai.helper.AndroidHelper;
 import org.fossasia.susi.ai.helper.Constant;
 import org.fossasia.susi.ai.helper.ConstraintsHelper;
-import org.fossasia.susi.ai.helper.MapHelper;
-import org.fossasia.susi.ai.helper.PrefManager;
 import org.fossasia.susi.ai.model.ChatMessage;
-import org.fossasia.susi.ai.model.MapData;
-import org.fossasia.susi.ai.model.WebLink;
-import org.fossasia.susi.ai.model.WebSearchModel;
-import org.fossasia.susi.ai.rest.clients.WebSearchClient;
-import org.fossasia.susi.ai.rest.services.WebSearchService;
-import org.fossasia.susi.ai.rest.responses.susi.Datum;
-import org.fossasia.susi.ai.rest.responses.others.WebSearch;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 import io.realm.OrderedRealmCollection;
 import io.realm.Realm;
 import io.realm.RealmChangeListener;
-import io.realm.RealmList;
 import io.realm.RealmResults;
 import pl.tajchert.sample.DotsTextView;
-import retrofit2.Call;
-import retrofit2.Callback;
-import retrofit2.Response;
 
 /**
  * <h1>Main adapter to display chat feed as a recycler view.</h1>
@@ -103,16 +69,15 @@ public class ChatFeedRecyclerAdapter extends SelectableAdapter implements Messag
     private static final int DATE_VIEW = 12;
     public int highlightMessagePosition = -1;
     public String query = "";
-    private String webquery ;
     private Context currContext;
     private Realm realm;
     private int lastMsgCount;
     private String TAG = ChatFeedRecyclerAdapter.class.getSimpleName();
     private RecyclerView recyclerView;
     private MessageViewHolder.ClickListener clickListener;
-    private ActionModeCallback actionModeCallback = new ActionModeCallback();
-    private ActionMode actionMode;
-    private SparseBooleanArray selectedItems;
+    public ActionModeCallback actionModeCallback = new ActionModeCallback();
+    public ActionMode actionMode;
+    public SparseBooleanArray selectedItems;
     private AppCompatActivity currActivity;
     // For typing dots from Susi
     private TypingDotsHolder dotsHolder;
@@ -162,7 +127,7 @@ public class ChatFeedRecyclerAdapter extends SelectableAdapter implements Messag
      * @param text String text
      * @return List of urls
      */
-    private static List<String> extractLinks(String text) {
+    public static List<String> extractLinks(String text) {
         List<String> links = new ArrayList<>();
         Matcher m = Patterns.WEB_URL.matcher(text);
         while (m.find()) {
@@ -220,7 +185,7 @@ public class ChatFeedRecyclerAdapter extends SelectableAdapter implements Messag
                 return new ChatViewHolder(view, clickListener, SUSI_IMAGE);
             case MAP:
                 view = inflater.inflate(R.layout.item_susi_map, viewGroup, false);
-                return new MapViewHolder(view, clickListener);
+                return new MapViewHolder(view);
             case USER_WITHLINK:
                 view = inflater.inflate(R.layout.item_user_link_preview, viewGroup, false);
                 return new LinkPreviewViewHolder(view, clickListener);
@@ -232,14 +197,14 @@ public class ChatFeedRecyclerAdapter extends SelectableAdapter implements Messag
                 return new PieChartViewHolder(view, clickListener);
             case SEARCH_RESULT:
                 view = inflater.inflate(R.layout.search_list, viewGroup, false);
-                SearchResultsListHolder searchResultsListHolder = new SearchResultsListHolder(view, clickListener);
+                SearchResultsListHolder searchResultsListHolder = new SearchResultsListHolder(view);
                 searchResultsListHolder.recyclerView.addItemDecoration(new ConstraintsHelper(6, currContext));
-                return new SearchResultsListHolder(view, clickListener);
+                return searchResultsListHolder;
             case WEB_SEARCH:
                 view = inflater.inflate(R.layout.search_list, viewGroup, false);
-                SearchResultsListHolder webResultsListHolder = new SearchResultsListHolder(view, clickListener);
+                SearchResultsListHolder webResultsListHolder = new SearchResultsListHolder(view);
                 webResultsListHolder.recyclerView.addItemDecoration(new ConstraintsHelper(6, currContext));
-                return new SearchResultsListHolder(view, clickListener);
+                return webResultsListHolder;
             case DATE_VIEW:
                 view = inflater.inflate(R.layout.date_view,viewGroup, false);
                 return new DateViewHolder(view);
@@ -350,126 +315,7 @@ public class ChatFeedRecyclerAdapter extends SelectableAdapter implements Messag
      */
     private void handleItemEvents(final SearchResultsListHolder searchResultsListHolder,final int position, boolean isClientSearch) {
         searchResultsListHolder.backgroundLayout.setBackgroundColor(ContextCompat.getColor(currContext, isSelected(position) ? R.color.translucent_blue : android.R.color.transparent));
-        final ChatMessage model = getData().get(position);
-        if(isClientSearch) {
-            if(model!=null) {
-                webquery = model.getWebquery();
-
-                if(model.getWebSearchList()==null || model.getWebSearchList().size()==0) {
-                    final WebSearchService apiService = WebSearchClient.getClient().create(WebSearchService.class);
-
-                    Call<WebSearch> call = apiService.getresult(webquery);
-
-                    call.enqueue(new Callback<WebSearch>() {
-                        @Override
-                        public void onResponse(Call<WebSearch> call, Response<WebSearch> response) {
-                            Log.e(TAG, response.toString());
-                            if (response.body() != null ) {
-                                realm.beginTransaction();
-                                RealmList<WebSearchModel> searchResults = new RealmList<>();
-
-                                for(int i=0 ; i<response.body().getRelatedTopics().size() ; i++) {
-                                    try {
-                                        String url = response.body().getRelatedTopics().get(i).getUrl();
-                                        String text = response.body().getRelatedTopics().get(i).getText();
-                                        String iconUrl = response.body().getRelatedTopics().get(i).getIcon().getUrl();
-                                        String htmlText = response.body().getRelatedTopics().get(i).getResult();
-                                        Realm realm = Realm.getDefaultInstance();
-                                        final WebSearchModel webSearch = realm.createObject(WebSearchModel.class);
-                                        try {
-                                            String[] tempStr = htmlText.split("\">");
-                                            String[] tempStr2 = tempStr[tempStr.length-1].split("</a>");
-                                            webSearch.setHeadline(tempStr2[0]);
-                                            webSearch.setBody(tempStr2[tempStr2.length-1]);
-                                        } catch (Exception e ) {
-                                            webSearch.setBody(text);
-                                            webSearch.setHeadline(webquery);
-                                        }
-                                        webSearch.setImageURL(iconUrl);
-                                        webSearch.setUrl(url);
-                                        searchResults.add(webSearch);
-                                    } catch (Exception e) {
-                                        Log.v(TAG,e.getLocalizedMessage());
-                                    }
-                                }
-                                if(searchResults.size()==0) {
-                                    Realm realm = Realm.getDefaultInstance();
-                                    final WebSearchModel webSearch = realm.createObject(WebSearchModel.class);
-                                    webSearch.setBody(null);
-                                    webSearch.setHeadline("No Results Found");
-                                    webSearch.setImageURL(null);
-                                    webSearch.setUrl(null);
-                                    searchResults.add(webSearch);
-                                }
-                                LinearLayoutManager layoutManager = new LinearLayoutManager(currContext,
-                                        LinearLayoutManager.HORIZONTAL, false);
-
-                                searchResultsListHolder.recyclerView.setLayoutManager(layoutManager);
-                                WebSearchAdapter resultsAdapter = new WebSearchAdapter(currContext, searchResults);
-                                searchResultsListHolder.recyclerView.setAdapter(resultsAdapter);
-                                model.setWebSearchList(searchResults);
-                                realm.copyToRealmOrUpdate(model);
-                                realm.commitTransaction();
-                            } else {
-                                searchResultsListHolder.recyclerView.setAdapter(null);
-                                searchResultsListHolder.recyclerView.setLayoutManager(null);
-                            }
-                        }
-
-                        @Override
-                        public void onFailure(Call<WebSearch> call, Throwable t) {
-                            Log.e(TAG, "error" + t.toString());
-                        }
-                    });
-                } else {
-                    LinearLayoutManager layoutManager = new LinearLayoutManager(currContext,
-                            LinearLayoutManager.HORIZONTAL, false);
-                    searchResultsListHolder.recyclerView.setLayoutManager(layoutManager);
-                    WebSearchAdapter resultsAdapter = new WebSearchAdapter(currContext, model.getWebSearchList());
-                    searchResultsListHolder.recyclerView.setAdapter(resultsAdapter);
-                }
-
-            } else{
-                searchResultsListHolder.recyclerView.setAdapter(null);
-                searchResultsListHolder.recyclerView.setLayoutManager(null);
-            }
-        } else {
-            if (model != null) {
-                LinearLayoutManager layoutManager = new LinearLayoutManager(currContext,
-                        LinearLayoutManager.HORIZONTAL, false);
-                searchResultsListHolder.recyclerView.setLayoutManager(layoutManager);
-                SearchResultsAdapter resultsAdapter;
-                List<Datum> data = new ArrayList<>();
-                int count = model.getCount();
-                if(count == -1) {
-                    resultsAdapter = new SearchResultsAdapter(currContext, model.getDatumRealmList());
-
-                } else {
-                    for (int i=0;i<count;i++) {
-                        data.add(model.getDatumRealmList().get(i));
-                    }
-                    resultsAdapter = new SearchResultsAdapter(currContext, data);
-                }
-                searchResultsListHolder.recyclerView.setAdapter(resultsAdapter);
-            } else {
-                searchResultsListHolder.recyclerView.setAdapter(null);
-                searchResultsListHolder.recyclerView.setLayoutManager(null);
-            }
-        }
-
-        searchResultsListHolder.backgroundLayout.setOnLongClickListener(new View.OnLongClickListener() {
-            @Override
-            public boolean onLongClick(View view) {
-                if (actionMode == null) {
-                    actionMode = ((AppCompatActivity) currContext).startSupportActionMode(actionModeCallback);
-                }
-
-                toggleSelectedItem(position);
-
-                return true;
-            }
-        });
-
+        searchResultsListHolder.setView(getData().get(position), isClientSearch, currContext);
     }
 
     /**
@@ -479,72 +325,11 @@ public class ChatFeedRecyclerAdapter extends SelectableAdapter implements Messag
      * @param position position of view
      */
     private void handleItemEvents(final ChatViewHolder chatViewHolder, final int position) {
-        final ChatMessage model = getData().get(position);
-
+        boolean flag = false;
+        if(highlightMessagePosition == position)
+            flag = true;
+        chatViewHolder.setView(getData().get(position), getItemViewType(position), flag, query);
         chatViewHolder.backgroundLayout.setBackgroundColor(ContextCompat.getColor(currContext, isSelected(position) ? R.color.translucent_blue : android.R.color.transparent));
-        if (model != null) {
-            try {
-                switch (getItemViewType(position)) {
-                    case USER_MESSAGE:
-                        chatViewHolder.chatTextView.setText(model.getContent());
-                        chatViewHolder.timeStamp.setText(model.getTimeStamp());
-                        if(model.getIsDelivered())
-                            chatViewHolder.receivedTick.setImageResource(R.drawable.ic_check);
-                        else
-                            chatViewHolder.receivedTick.setImageResource(R.drawable.ic_clock);
-
-                        chatViewHolder.chatTextView.setTag(chatViewHolder);
-                        if (highlightMessagePosition == position) {
-                            String text = chatViewHolder.chatTextView.getText().toString();
-                            SpannableString modify = new SpannableString(text);
-                            Pattern pattern = Pattern.compile(query, Pattern.CASE_INSENSITIVE);
-                            Matcher matcher = pattern.matcher(modify);
-                            while (matcher.find()) {
-                                int startIndex = matcher.start();
-                                int endIndex = matcher.end();
-                                modify.setSpan(new BackgroundColorSpan(Color.parseColor("#ffff00")), startIndex, endIndex, Spannable.SPAN_EXCLUSIVE_EXCLUSIVE);
-                            }
-                            chatViewHolder.chatTextView.setText(modify);
-
-                        }
-                        chatViewHolder.timeStamp.setTag(chatViewHolder);
-                        chatViewHolder.receivedTick.setTag(chatViewHolder);
-                        break;
-                    case SUSI_MESSAGE:
-                        Spanned answerText;
-                        if (android.os.Build.VERSION.SDK_INT >= Build.VERSION_CODES.N){
-                            answerText = Html.fromHtml(model.getContent(),Html.FROM_HTML_MODE_COMPACT);
-                        } else{
-                            answerText = Html.fromHtml(model.getContent());
-                        }
-
-                        chatViewHolder.chatTextView.setText(answerText);
-                        chatViewHolder.chatTextView.setMovementMethod(LinkMovementMethod.getInstance());
-                        chatViewHolder.timeStamp.setText(model.getTimeStamp());
-                        chatViewHolder.chatTextView.setTag(chatViewHolder);
-                        if (highlightMessagePosition == position) {
-                            String text = chatViewHolder.chatTextView.getText().toString();
-                            SpannableString modify = new SpannableString(text);
-                            Pattern pattern = Pattern.compile(query, Pattern.CASE_INSENSITIVE);
-                            Matcher matcher = pattern.matcher(modify);
-                            while (matcher.find()) {
-                                int startIndex = matcher.start();
-                                int endIndex = matcher.end();
-                                modify.setSpan(new BackgroundColorSpan(Color.parseColor("#ffff00")), startIndex, endIndex, Spannable.SPAN_EXCLUSIVE_EXCLUSIVE);
-                            }
-                            chatViewHolder.chatTextView.setText(modify);
-
-                        }
-                        chatViewHolder.timeStamp.setTag(chatViewHolder);
-                        break;
-                    case USER_IMAGE:
-                    case SUSI_IMAGE:
-                    default:
-                }
-            } catch (Exception e) {
-                e.printStackTrace();
-            }
-        }
     }
 
     /**
@@ -554,55 +339,8 @@ public class ChatFeedRecyclerAdapter extends SelectableAdapter implements Messag
      * @param position position of view
      */
     private void handleItemEvents(final MapViewHolder mapViewHolder, final int position) {
-        final ChatMessage model = getData().get(position);
+        mapViewHolder.setView(getData().get(position), currContext, this, position);
         mapViewHolder.backgroundLayout.setBackgroundColor(ContextCompat.getColor(currContext, isSelected(position) ? R.color.translucent_blue : android.R.color.transparent));
-
-        if (model != null) {
-            try {
-                final MapHelper mapHelper = new MapHelper(new MapData(model.getLatitude(),model.getLongitude(),model.getZoom()));
-                mapViewHolder.pointer.setVisibility(View.GONE);
-                Log.v(TAG, mapHelper.getMapURL());
-
-                Picasso.with(currContext).load(mapHelper.getMapURL())
-                        .into(mapViewHolder.mapImage, new com.squareup.picasso.Callback() {
-                            @Override
-                            public void onSuccess() {
-                                mapViewHolder.pointer.setVisibility(View.VISIBLE);
-                            }
-
-                            @Override
-                            public void onError() {
-                                Log.d("Error","map image can't loaded");
-                            }
-                        });
-
-                mapViewHolder.mapImage.setOnClickListener(new View.OnClickListener() {
-                    @Override
-                    public void onClick(View v) {
-                        /*
-                          Open in Google Maps if installed, otherwise open browser.
-                        */
-                        if(selectedItems.size() == 0) {
-                            Intent mapIntent;
-                            if (AndroidHelper.isGoogleMapsInstalled(currContext) && mapHelper.isParseSuccessful()) {
-                                Uri gmmIntentUri = Uri.parse(String.format("geo:%s,%s?z=%s", model.getLatitude(), model.getLongitude(), model.getZoom()));
-                                mapIntent = new Intent(Intent.ACTION_VIEW, gmmIntentUri);
-                                mapIntent.setPackage(AndroidHelper.GOOGLE_MAPS_PKG);
-                            } else {
-                                mapIntent = new Intent(Intent.ACTION_VIEW);
-                                mapIntent.setData(Uri.parse(mapHelper.getWebLink()));
-                            }
-                            currContext.startActivity(mapIntent);
-                        } else {
-                            toggleSelectedItem(position);
-                        }
-                    }
-                });
-
-            } catch (Exception e) {
-                e.printStackTrace();
-            }
-        }
     }
 
     /**
@@ -613,218 +351,8 @@ public class ChatFeedRecyclerAdapter extends SelectableAdapter implements Messag
      */
     private void handleItemEvents(final LinkPreviewViewHolder linkPreviewViewHolder, final int position) {
         final ChatMessage model = getData().get(position);
-        linkPreviewViewHolder.text.setText(model.getContent());
-        linkPreviewViewHolder.timestampTextView.setText(model.getTimeStamp());
+        linkPreviewViewHolder.setView(getData().get(position), currContext, this, highlightMessagePosition, position, query);
         linkPreviewViewHolder.backgroundLayout.setBackgroundColor(ContextCompat.getColor(currContext, isSelected(position) ? R.color.translucent_blue : android.R.color.transparent));
-
-        if (model.getWebLinkData() == null) {
-            LinkPreviewCallback linkPreviewCallback = new LinkPreviewCallback() {
-                @Override
-                public void onPre() {
-                    linkPreviewViewHolder.previewImageView.setVisibility(View.GONE);
-                    linkPreviewViewHolder.descriptionTextView.setVisibility(View.GONE);
-                    linkPreviewViewHolder.titleTextView.setVisibility(View.GONE);
-                    linkPreviewViewHolder.previewLayout.setVisibility(View.GONE);
-                }
-
-                @Override
-                public void onPos(final SourceContent sourceContent, boolean b) {
-                    if(!PrefManager.hasTokenExpired()) {
-                        realm.beginTransaction();
-                        Realm realm = Realm.getDefaultInstance();
-                        WebLink link = realm.createObject(WebLink.class);
-
-                        if (sourceContent != null) {
-
-                            if (!sourceContent.getDescription().isEmpty()) {
-                                Log.d(TAG, "onPos: " + sourceContent.getDescription());
-                                linkPreviewViewHolder.previewLayout.setVisibility(View.VISIBLE);
-                                linkPreviewViewHolder.descriptionTextView.setVisibility(View.VISIBLE);
-                                linkPreviewViewHolder.descriptionTextView.setText(sourceContent.getDescription());
-                            }
-
-                            if (!sourceContent.getTitle().isEmpty()) {
-                                Log.d(TAG, "onPos: " + sourceContent.getTitle());
-                                linkPreviewViewHolder.previewLayout.setVisibility(View.VISIBLE);
-                                linkPreviewViewHolder.titleTextView.setVisibility(View.VISIBLE);
-                                linkPreviewViewHolder.titleTextView.setText(sourceContent.getTitle());
-                            }
-
-                            link.setBody(sourceContent.getDescription());
-                            link.setHeadline(sourceContent.getTitle());
-                            link.setUrl(sourceContent.getUrl());
-
-                            final List<String> imageList = sourceContent.getImages();
-
-                            if (imageList == null || imageList.size() == 0) {
-                                linkPreviewViewHolder.previewImageView.setVisibility(View.GONE);
-                                link.setImageURL("");
-                            } else {
-                                linkPreviewViewHolder.previewImageView.setVisibility(View.VISIBLE);
-                                Picasso.with(currContext).load(imageList.get(0))
-                                        .fit().centerCrop()
-                                        .into(linkPreviewViewHolder.previewImageView);
-                                link.setImageURL(imageList.get(0));
-                            }
-
-                            linkPreviewViewHolder.previewLayout.setOnClickListener(new View.OnClickListener() {
-                                @Override
-                                public void onClick(View view) {
-                                    if (selectedItems.size() == 0) {
-                                        Uri webpage = Uri.parse(sourceContent.getFinalUrl());
-                                        Intent intent = new Intent(Intent.ACTION_VIEW, webpage);
-                                        if (intent.resolveActivity(currContext.getPackageManager()) != null) {
-                                            currContext.startActivity(intent);
-                                        }
-                                    } else {
-                                        toggleSelectedItem(position);
-                                    }
-                                }
-                            });
-
-                            linkPreviewViewHolder.text.setOnClickListener(new View.OnClickListener() {
-                                @Override
-                                public void onClick(View view) {
-                                    if (selectedItems.size() != 0)
-                                        toggleSelectedItem(position);
-                                }
-                            });
-                        }
-
-                        linkPreviewViewHolder.previewLayout.setOnLongClickListener(new View.OnLongClickListener() {
-                            @Override
-                            public boolean onLongClick(View view) {
-                                if (actionMode == null) {
-                                    actionMode = ((AppCompatActivity) currContext).startSupportActionMode(actionModeCallback);
-                                }
-                                toggleSelectedItem(position);
-
-                                return true;
-                            }
-                        });
-
-                        linkPreviewViewHolder.text.setOnLongClickListener(new View.OnLongClickListener() {
-                            @Override
-                            public boolean onLongClick(View view) {
-                                if (actionMode == null) {
-                                    actionMode = ((AppCompatActivity) currContext).startSupportActionMode(actionModeCallback);
-                                }
-
-                                toggleSelectedItem(position);
-
-                                return true;
-                            }
-                        });
-
-                        model.setWebLinkData(link);
-                        realm.copyToRealmOrUpdate(model);
-                        realm.commitTransaction();
-                    }
-                }
-            };
-
-            if (model != null) {
-                List<String> urlList = extractLinks(model.getContent());
-                StringBuilder url = new StringBuilder(urlList.get(0));
-                StringBuilder http = new StringBuilder("http://");
-                StringBuilder https = new StringBuilder("https://");
-                if (!(url.toString().startsWith(http.toString()) || url.toString().startsWith(https.toString()))) {
-                    url = http.append(url.toString());
-                }
-                TextCrawler textCrawler = new TextCrawler();
-                textCrawler.makePreview(linkPreviewCallback, url.toString());
-            }
-        } else {
-
-            if(!model.getWebLinkData().getHeadline().isEmpty()) {
-                Log.d(TAG, "onPos: " + model.getWebLinkData().getHeadline());
-                linkPreviewViewHolder.titleTextView.setText(model.getWebLinkData().getHeadline());
-            } else {
-                linkPreviewViewHolder.titleTextView.setVisibility(View.GONE);
-                Log.d(TAG, "handleItemEvents: " + "isEmpty");
-            }
-
-            if(!model.getWebLinkData().getBody().isEmpty()) {
-                Log.d(TAG, "onPos: " + model.getWebLinkData().getHeadline());
-                linkPreviewViewHolder.descriptionTextView.setText(model.getWebLinkData().getBody());
-            } else {
-                linkPreviewViewHolder.descriptionTextView.setVisibility(View.GONE);
-                Log.d(TAG, "handleItemEvents: " + "isEmpty");
-            }
-
-            if(model.getWebLinkData().getHeadline().isEmpty() && model.getWebLinkData().getBody().isEmpty()) {
-                linkPreviewViewHolder.previewLayout.setVisibility(View.GONE);
-            }
-
-            Log.i(TAG, model.getWebLinkData().getImageURL());
-            if (!model.getWebLinkData().getImageURL().equals("")) {
-                Picasso.with(currContext).load(model.getWebLinkData().getImageURL())
-                        .fit().centerCrop()
-                        .into(linkPreviewViewHolder.previewImageView);
-            } else {
-                linkPreviewViewHolder.previewImageView.setVisibility(View.GONE);
-            }
-
-            linkPreviewViewHolder.previewLayout.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View view) {
-                    if(selectedItems.size() == 0) {
-                        Uri webpage = Uri.parse(model.getWebLinkData().getUrl());
-                        Intent intent = new Intent(Intent.ACTION_VIEW, webpage);
-                        if (intent.resolveActivity(currContext.getPackageManager()) != null) {
-                            currContext.startActivity(intent);
-                        }
-                    } else {
-                        toggleSelectedItem(position);
-                    }
-                }
-            });
-
-            linkPreviewViewHolder.text.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View view) {
-                    if(selectedItems.size() != 0)
-                        toggleSelectedItem(position);
-                }
-            });
-
-            linkPreviewViewHolder.previewLayout.setOnLongClickListener(new View.OnLongClickListener() {
-                @Override
-                public boolean onLongClick(View view) {
-                    if (actionMode == null) {
-                        actionMode = ((AppCompatActivity) currContext).startSupportActionMode(actionModeCallback);
-                    }
-                    toggleSelectedItem(position);
-
-                    return true;
-                }
-            });
-
-            linkPreviewViewHolder.text.setOnLongClickListener(new View.OnLongClickListener() {
-                @Override
-                public boolean onLongClick(View view) {
-                    if (actionMode == null) {
-                        actionMode = ((AppCompatActivity) currContext).startSupportActionMode(actionModeCallback);
-                    }
-                    toggleSelectedItem(position);
-
-                    return true;
-                }
-            });
-        }
-
-        if (highlightMessagePosition == position) {
-            String text = linkPreviewViewHolder.text.getText().toString();
-            SpannableString modify = new SpannableString(text);
-            Pattern pattern = Pattern.compile(query, Pattern.CASE_INSENSITIVE);
-            Matcher matcher = pattern.matcher(modify);
-            while (matcher.find()) {
-                int startIndex = matcher.start();
-                int endIndex = matcher.end();
-                modify.setSpan(new BackgroundColorSpan(Color.parseColor("#ffff00")), startIndex, endIndex, Spannable.SPAN_EXCLUSIVE_EXCLUSIVE);
-            }
-            linkPreviewViewHolder.text.setText(modify);
-        }
         if (getItemViewType(position) == USER_WITHLINK) {
             if (model.getIsDelivered())
                 linkPreviewViewHolder.receivedTick.setImageResource(R.drawable.ic_check);
@@ -840,56 +368,8 @@ public class ChatFeedRecyclerAdapter extends SelectableAdapter implements Messag
      * @param position position of view
      */
     private void handleItemEvents(final PieChartViewHolder pieChartViewHolder, final int position) {
-        final ChatMessage model = getData().get(position);
-        if (model != null) {
-            try {
-                pieChartViewHolder.chatTextView.setText(model.getContent());
-                pieChartViewHolder.backgroundLayout.setBackgroundColor(ContextCompat.getColor(currContext, isSelected(position) ? R.color.translucent_blue : android.R.color.transparent));
-                pieChartViewHolder.timeStamp.setText(model.getTimeStamp());
-                pieChartViewHolder.pieChart.setUsePercentValues(true);
-                pieChartViewHolder.pieChart.setDrawHoleEnabled(true);
-                pieChartViewHolder.pieChart.setHoleRadius(7);
-                pieChartViewHolder.pieChart.setTransparentCircleRadius(10);
-                pieChartViewHolder.pieChart.setRotationEnabled(true);
-                pieChartViewHolder.pieChart.setRotationAngle(0);
-                pieChartViewHolder.pieChart.setDragDecelerationFrictionCoef(0.001f);
-                pieChartViewHolder.pieChart.getLegend().setEnabled(false);
-                pieChartViewHolder.pieChart.setDescription("");
-                RealmList<Datum> datumList = model.getDatumRealmList();
-                final ArrayList<Entry> yVals = new ArrayList<>();
-                final ArrayList<String> xVals = new ArrayList<>();
-                for (int i = 0; i < datumList.size(); i++) {
-                    yVals.add(new Entry(datumList.get(i).getPercent(), i));
-                    xVals.add(datumList.get(i).getPresident());
-                }
-                pieChartViewHolder.pieChart.setClickable(false);
-                pieChartViewHolder.pieChart.setHighlightPerTapEnabled(false);
-                PieDataSet dataSet = new PieDataSet(yVals, "");
-                dataSet.setSliceSpace(3);
-                dataSet.setSelectionShift(5);
-                ArrayList<Integer> colors = new ArrayList<>();
-                for (int c : ColorTemplate.VORDIPLOM_COLORS)
-                    colors.add(c);
-                for (int c : ColorTemplate.JOYFUL_COLORS)
-                    colors.add(c);
-                for (int c : ColorTemplate.COLORFUL_COLORS)
-                    colors.add(c);
-                for (int c : ColorTemplate.LIBERTY_COLORS)
-                    colors.add(c);
-                for (int c : ColorTemplate.PASTEL_COLORS)
-                    colors.add(c);
-                dataSet.setColors(colors);
-                PieData data = new PieData(xVals, dataSet);
-                data.setValueFormatter(new PercentFormatter());
-                data.setValueTextSize(11f);
-                data.setValueTextColor(Color.GRAY);
-                pieChartViewHolder.pieChart.setData(data);
-                pieChartViewHolder.pieChart.highlightValues(null);
-                pieChartViewHolder.pieChart.invalidate();
-            } catch (Exception e) {
-                e.printStackTrace();
-            }
-        }
+        pieChartViewHolder.setView(getData().get(position));
+        pieChartViewHolder.backgroundLayout.setBackgroundColor(ContextCompat.getColor(currContext, isSelected(position) ? R.color.translucent_blue : android.R.color.transparent));
     }
 
     /**
@@ -917,7 +397,7 @@ public class ChatFeedRecyclerAdapter extends SelectableAdapter implements Messag
      *
      * @param position position of message
      */
-    private void  toggleSelectedItem(int position) {
+    public void  toggleSelectedItem(int position) {
         toggleSelection(position);
         int count = getSelectedItemCount();
 
@@ -973,26 +453,8 @@ public class ChatFeedRecyclerAdapter extends SelectableAdapter implements Messag
         @Override
         public boolean onPrepareActionMode(ActionMode mode, Menu menu) {
 
-            if(getSelectedItems().size()>0) {
-                Log.d(TAG, "onPrepareActionMode: size " + getSelectedItems().size());
-                for(int i = 0; i<getSelectedItems().size() ;i++ ){
-                    if(getItemViewType(getSelectedItems().get(i)) == USER_MESSAGE ||
-                            getItemViewType(getSelectedItems().get(i)) == SUSI_MESSAGE ||
-                            getItemViewType(getSelectedItems().get(i)) == USER_WITHLINK ||
-                            getItemViewType(getSelectedItems().get(i)) == SUSI_WITHLINK){
-                        menu.clear();
-                        mode.getMenuInflater().inflate(R.menu.menu_selection_mode, menu);
-                    }
-                    else {
-                        Log.d(TAG, "onPrepareActionMode: + Other response");
-                        menu.removeItem(R.id.menu_item_copy);
-                        menu.removeItem(R.id.menu_item_share);
-                        break;
-                    }
-                }
-                Log.d(TAG, "onPrepareActionMode: " + getItemViewType(getSelectedItems().get(0)));
-            }
-
+            menu.clear();
+            mode.getMenuInflater().inflate(R.menu.menu_selection_mode, menu);
             return false;
         }
