+++ app/src/main/java/org/fossasia/susi/ai/adapters/viewholders/ChatViewHolder.java
@@ -1,12 +1,25 @@
 package org.fossasia.susi.ai.adapters.viewholders;
 
+import android.graphics.Color;
+import android.os.Build;
 import android.support.annotation.Nullable;
+import android.text.Html;
+import android.text.Spannable;
+import android.text.SpannableString;
+import android.text.Spanned;
+import android.text.method.LinkMovementMethod;
+import android.text.style.BackgroundColorSpan;
 import android.view.View;
 import android.widget.ImageView;
 import android.widget.LinearLayout;
 import android.widget.TextView;
 
 import org.fossasia.susi.ai.R;
+import org.fossasia.susi.ai.helper.Constant;
+import org.fossasia.susi.ai.model.ChatMessage;
+
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 import butterknife.BindView;
 import butterknife.ButterKnife;
@@ -24,7 +37,7 @@ import static org.fossasia.susi.ai.adapters.recycleradapters.ChatFeedRecyclerAda
  * --25/09/16 at
  * --9:51 PM
  */
-public class ChatViewHolder extends MessageViewHolder{
+public class ChatViewHolder extends MessageViewHolder {
 
     @BindView(R.id.text)
     public TextView chatTextView;
@@ -32,8 +45,10 @@ public class ChatViewHolder extends MessageViewHolder{
     public TextView timeStamp;
     @BindView(R.id.background_layout)
     public LinearLayout backgroundLayout;
-    @Nullable @BindView(R.id.received_tick)
+    @Nullable
+    @BindView(R.id.received_tick)
     public ImageView receivedTick;
+
 
     /**
      * Instantiates a new Chat view holder.
@@ -42,8 +57,8 @@ public class ChatViewHolder extends MessageViewHolder{
      * @param clickListener the click listener
      * @param myMessage     the my message
      */
-    public ChatViewHolder(View view, ClickListener clickListener ,int myMessage) {
-        super(view,clickListener);
+    public ChatViewHolder(View view, ClickListener clickListener, int myMessage) {
+        super(view, clickListener);
         ButterKnife.bind(this, view);
         switch (myMessage) {
             case USER_MESSAGE:
@@ -53,6 +68,85 @@ public class ChatViewHolder extends MessageViewHolder{
             case USER_IMAGE:
             case SUSI_IMAGE:
             default:
+        }
+    }
+
+    /**
+     * Inflate ChatView
+     *
+     * @param model the ChatMessage object
+     * @param viewType the viewType
+     * @param flag the flag
+     * @param query the query
+     */
+    public void setView(ChatMessage model, int viewType, boolean flag, String query) {
+        if (model != null) {
+            try {
+                switch (viewType) {
+                    case USER_MESSAGE:
+                        chatTextView.setText(model.getContent());
+                        timeStamp.setText(model.getTimeStamp());
+                        if (model.getIsDelivered())
+                            receivedTick.setImageResource(R.drawable.ic_check);
+                        else
+                            receivedTick.setImageResource(R.drawable.ic_clock);
+
+                        chatTextView.setTag(this);
+                        if (flag) {
+                            String text = chatTextView.getText().toString();
+                            SpannableString modify = new SpannableString(text);
+                            Pattern pattern = Pattern.compile(query, Pattern.CASE_INSENSITIVE);
+                            Matcher matcher = pattern.matcher(modify);
+                            while (matcher.find()) {
+                                int startIndex = matcher.start();
+                                int endIndex = matcher.end();
+                                modify.setSpan(new BackgroundColorSpan(Color.parseColor("#ffff00")), startIndex, endIndex, Spannable.SPAN_EXCLUSIVE_EXCLUSIVE);
+                            }
+                            chatTextView.setText(modify);
+
+                        }
+                        timeStamp.setTag(this);
+                        receivedTick.setTag(this);
+                        break;
+                    case SUSI_MESSAGE:
+                        Spanned answerText;
+                        if(model.getActionType().equals(Constant.ANCHOR)) {
+                            chatTextView.setLinksClickable(true);
+                            chatTextView.setMovementMethod(LinkMovementMethod.getInstance());
+                        } else {
+                            chatTextView.setLinksClickable(false);
+                        }
+                        if (android.os.Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
+                            answerText = Html.fromHtml(model.getContent(), Html.FROM_HTML_MODE_COMPACT);
+                        } else {
+                            answerText = Html.fromHtml(model.getContent());
+                        }
+
+                        chatTextView.setText(answerText);
+                        timeStamp.setText(model.getTimeStamp());
+                        chatTextView.setTag(this);
+                        if (flag) {
+                            String text = chatTextView.getText().toString();
+                            SpannableString modify = new SpannableString(text);
+                            Pattern pattern = Pattern.compile(query, Pattern.CASE_INSENSITIVE);
+                            Matcher matcher = pattern.matcher(modify);
+                            while (matcher.find()) {
+                                int startIndex = matcher.start();
+                                int endIndex = matcher.end();
+                                modify.setSpan(new BackgroundColorSpan(Color.parseColor("#ffff00")), startIndex, endIndex, Spannable.SPAN_EXCLUSIVE_EXCLUSIVE);
+                            }
+                            chatTextView.setText(modify);
+
+                        }
+                        timeStamp.setTag(this);
+                        break;
+                    case USER_IMAGE:
+                    case SUSI_IMAGE:
+                    default:
+                }
+            } catch (Exception e) {
+                e.printStackTrace();
+            }
         }
     }
 }
